+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.1.24 (unreleased)
 -------------------
 
+- WEB-3984 : Remove deprecated cropping annotations on banner
+  [boulch, laulaz]
+
 - WEB-3984 : Don't get banner scale anymore. Get full banner image directly
   [boulch, laulaz]